Fixed: Installer not copying files and launching application

+++ Comet/Controls/ProgressDialog.cs
@@ -278,7 +278,11 @@
         /// </summary>
         private void InstallUpdate()
         {
-            Directory.Delete(_installOptions.InstallFilesFolder, true);
+            if (Directory.Exists(_installOptions.InstallFilesFolder))
+            {
+                Directory.Delete(_installOptions.InstallFilesFolder, true);
+            }
+
             FileManager.CreateDirectory(_installOptions.InstallFilesFolder);
 
             foreach (string _archive in _installOptions.DownloadedFiles)
